refactor(github): delete Check projection repository
